限制速度可用

+++ shadowsocks/config.go
@@ -4,7 +4,7 @@ type SSconfig struct {
 	Uid        int    `json:"uid"`
 	Sid        int    `json:"sid"`
 	Timeout    int    `json:"timeout"`
-	Limit      int64  `json:"limit"`
+	Limit      int    `json:"limit"`
 	Method     string `json:"method"`
 	Password   string `json:"password"`
 	Expiration int64  `json:"expiration"`
